refactor: refactor scraper service
Refactor the data fetching component out of the parsing classes and move it to the scraper itself in order to reduce duplicate code and encourage single responsability.

Refactor parsers to take an html content string as a parameter instead of a url.

Refactor scraper to output an html content string instead of a url. The scraper will extract the html content from that url and return that instead.